requirements.txt
streamlit
transformers
torch
ultralytics
opencv-python-headless
Pillow